requirements.txt
PyOpenGL>=3.1.7
PyGLM>=2.7.1
numpy>=1.26.4
